added token authentication

+++ app/controllers/api/v1/accounts_controller.rb
@@ -3,43 +3,38 @@ module Api
   module V1
     class AccountsController < ApplicationController
       # displays the data of all the accounts
-      api :GET, '/accounts/', 'Fetch all the account details'
+      api :GET, '/sessions/:session_id/accounts/', 'Fetch all the account details'
+      param :session_id, String, desc: 'Session Id'
       example %(
-      GET /api/v1/accounts
+      GET /api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/accounts
       {
       "status": "SUCCESS",
       "message": "Loaded accounts",
       "data": [
-        {
-          "id": 1,
-          "code": "PQR100",
-          "name": "Account1",
-          "password": "Account1",
-          "created_at": "2020-02-25T14:00:01.000Z",
-          "updated_at": "2020-02-25T14:08:55.000Z"
-        },
-        {
-          "id": 2,
-          "code": "XYZ100",
-          "name": "Account2",
-          "password": "Account2",
-          "created_at": "2020-02-25T14:00:19.000Z",
-          "updated_at": "2020-02-25T14:00:19.000Z"
-        }
-      ]
+      {
+        "id": 18,
+        "code": "Account11",
+        "name": "new-account",
+        "password": "Account11",
+        "created_at": "2020-03-14T12:36:37.000Z",
+        "updated_at": "2020-03-14T12:48:50.000Z",
+        "user_id": 1
+      }]
       }
       )
       def index
-        accounts = Account.all
+        current_user = User.find_by_authentication_token(params[:session_id])
+        accounts = current_user.accounts.all
         render json: { status: 'SUCCESS', message: 'Loaded accounts', data:
  accounts }, status: :ok
       end
 
       # displays the details of specific account based on the id
-      api :GET, '/accounts/:id', 'Fetches the specific account based on id'
+      api :GET, '/sessions/:session_id/accounts/:id', 'Fetches the specific account based on id'
+      param :session_id, String, desc: 'Session Id'
       param :id, :number, desc: 'Id'
       example %(
-      GET /api/v1/accounts/1
+      GET /api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/accounts/1
       {
       "status": "SUCCESS",
       "message": "Loaded account",
@@ -49,23 +44,27 @@ module Api
         "name": "Account1",
         "password": "Account1",
         "created_at": "2020-02-25T14:00:01.000Z",
-        "updated_at": "2020-02-25T14:08:55.000Z"
+        "updated_at": "2020-02-25T14:08:55.000Z",
+        "user_id": 1
       }
       }
       )
       def show
-        account = Account.find(params[:id])
+        current_user = User.find_by_authentication_token(params[:session_id])
+        account = current_user.accounts.find(params[:id])
+        #account = Account.find(params[:id])
         render json: { status: 'SUCCESS', message: 'Loaded account', data:
 account }, status: :ok
       end
 
       # creates a new account
-      api :POST, '/accounts/', 'Creates a new account'
+      api :POST, '/sessions/:session_id/accounts/', 'Creates a new account'
+      param :session_id, String, desc: 'Session Id'
       param :code, String, desc: 'Code', required: true
       param :name, String, desc: 'Name', required: true
       param :password, String, desc: 'Password', required: true
       example %(
-      POST /api/v1/accounts
+      POST /api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/accounts
       Input:
       {
       "code":"SDA100",
@@ -83,26 +82,31 @@ account }, status: :ok
         "name": "Account11",
         "password": "Account11",
         "created_at": "2020-03-02T10:24:09.000Z",
-        "updated_at": "2020-03-02T10:24:09.000Z"
+        "updated_at": "2020-03-02T10:24:09.000Z",
+        "user_id": 1
       }
       }
       )
       def create
-        account = Account.new(params_account)
-        if account.save
-          render json: { status: 'SUCCESS', message: 'Saved account', data:
-account }, status: :ok
-        else
-          render json: { status: 'ERROR', message: 'Account not saved', data:
-account.errors }, status: :unprocessable_entity
+        current_user = User.find_by_authentication_token(params[:session_id])
+        if current_user != nil
+          account = Account.new({name: params[:name], code: params[:code], password: params[:password], user_id: current_user.id})
+          if account.save
+            render json: { status: 'SUCCESS', message: 'Saved account', data:
+  account }, status: :ok
+          else
+            render json: { status: 'ERROR', message: 'Account not saved', data:
+  account.errors }, status: :unprocessable_entity
+          end
         end
       end
 
       # deletes a specific account based on the id
-      api :DELETE, '/accounts/:id', 'Deletes a specific account by id'
+      api :DELETE, '/sessions/:session_id/accounts/:id', 'Deletes a specific account by id'
+      param :session_id, String, desc: 'Session Id'
       param :id, :number, desc: 'Id'
       example %(
-      DELETE api/v1/accounts/14
+      DELETE api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/accounts/14
       {
       "status": "SUCCESS",
       "message": "Deleted account",
@@ -112,25 +116,30 @@ account.errors }, status: :unprocessable_entity
         "name": "Account11",
         "password": "Account11",
         "created_at": "2020-03-02T10:24:09.000Z",
-        "updated_at": "2020-03-02T10:24:09.000Z"
+        "updated_at": "2020-03-02T10:24:09.000Z",
+        "user_id": 1
       }
       }
       )
       def destroy
-        account = Account.find(params[:id])
-        account.destroy
-        render json: { status: 'SUCCESS', message: 'Deleted account', data:
-account }, status: :ok
+        current_user = User.find_by_authentication_token(params[:session_id])
+        if current_user != nil
+          account = Account.find(params[:id])
+          account.destroy
+          render json: { status: 'SUCCESS', message: 'Deleted account', data:
+  account }, status: :ok
+        end
       end
 
       # updates a specific account based on the id
-      api :PUT, '/accounts/:id', 'Updates a specfic account by id'
+      api :PUT, '/sessions/:session_id/accounts/:id', 'Updates a specfic account by id'
+      param :session_id, String, desc: 'Session Id'
       param :id, :number, desc: 'Id'
       param :code, String, desc: 'Code'
       param :name, String, desc: 'Name'
       param :password, String, desc: 'Password'
       example %(
-      PUT api/v1/accounts/13
+      PUT api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/accounts/13
       Input:
       {
       "code":"SDA100",
@@ -148,27 +157,32 @@ account }, status: :ok
         "name": "Account13",
         "password": "Account13",
         "created_at": "2020-02-28T09:15:52.000Z",
-        "updated_at": "2020-03-02T10:33:47.000Z"
+        "updated_at": "2020-03-02T10:33:47.000Z",
+        "user_id": 1
       }
       }
       )
       def update
-        account = Account.find(params[:id])
-        if account.update_attributes(params_account)
-          render json: { status: 'SUCCESS', message: 'Updated account', data:
-account }, status: :ok
-        else
-          render json: { status: 'ERROR', message: 'Account not updated', data:
-account.errors }, status: :unprocessable_entity
+        current_user = User.find_by_authentication_token(params[:session_id])
+        if current_user != nil
+          account = Account.find(params[:id])
+          if account.update_attributes(params_account)
+            render json: { status: 'SUCCESS', message: 'Updated account', data:
+  account }, status: :ok
+          else
+            render json: { status: 'ERROR', message: 'Account not updated', data:
+  account.errors }, status: :unprocessable_entity
+          end
         end
       end
 
       # displays medias(audios/videos) of specific account based on the id
-      api :GET, '/show_media/:id', 'Displays medias(audios or videos) of
+      api :GET, '/sessions/:session_id/show_media/:id', 'Displays medias(audios or videos) of
 specific account based on id'
+      param :session_id, String, desc: 'Session Id'
       param :id, :number, desc: 'Id'
       example %(
-      GET api/v1/show_media/2
+      GET api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/show_media/2
       {
       "status": "SUCCESS",
       "message": "Loaded media",
@@ -201,19 +215,23 @@ specific account based on id'
       }
       )
       def show_media
-        account = Account.find(params[:id])
-        medias = account.medias.all
-        render json: { status: 'SUCCESS', message: 'Loaded media', data: medias
-}, status: :ok
+        current_user = User.find_by_authentication_token(params[:session_id])
+        if current_user != nil
+          account = Account.find(params[:id])
+          medias = account.medias.all
+          render json: { status: 'SUCCESS', message: 'Loaded media', data: medias
+  }, status: :ok
+        end
       end
 
       # displays medias(audios/videos) of specific account based on the account
       # code
-      api :GET, '/show_media_by_code/:id', 'Displays medias(audios or videos) of
+      api :GET, '/sessions/:session_id/show_media_by_code/:id', 'Displays medias(audios or videos) of
 specific account based on code'
+      param :session_id, String, desc: 'Session Id'
       param :id, String, desc: 'Code'
       example %(
-      GET api/v1/show_media_by_code/XYZ100
+      GET api/v1/sessions/pvVSdt9Q2f9up4fa5AQV/show_media_by_code/XYZ100
       {
       "status": "SUCCESS",
       "message": "Loaded media",
@@ -246,10 +264,13 @@ specific account based on code'
       }
       )
       def show_media_by_code
-        account = Account.find_by_code(params[:id])
-        medias = account.medias.all
-        render json: { status: 'SUCCESS', message: 'Loaded media', data: medias
-}, status: :ok
+        current_user = User.find_by_authentication_token(params[:session_id])
+        if current_user != nil
+          account = Account.find_by_code(params[:id])
+          medias = account.medias.all
+          render json: { status: 'SUCCESS', message: 'Loaded media', data: medias
+  }, status: :ok
+        end
       end
 
       private
